+++ src/Endpoint/MyEnvelopesEndpoint.php
@@ -8,6 +8,7 @@ use DigitalCz\DigiSign\Endpoint\Traits\GetEndpointTrait;
 use DigitalCz\DigiSign\Endpoint\Traits\ListEndpointTrait;
 use DigitalCz\DigiSign\Resource\ListResource;
 use DigitalCz\DigiSign\Resource\MyEnvelope;
+use DigitalCz\DigiSign\Resource\MyEnvelopeInfo;
 
 /**
  * @extends ResourceEndpoint<MyEnvelope>
@@ -39,5 +40,10 @@ final class MyEnvelopesEndpoint extends ResourceEndpoint
     public function listWaitingForOthers(array $query = []): ListResource
     {
         return $this->makeListResource($this->getRequest('/waiting-for-others', ['query' => $query]));
+    }
+
+    public function info(string $id): MyEnvelopeInfo
+    {
+        return $this->createResource($this->getRequest('/{id}/info', ['id' => $id]), MyEnvelopeInfo::class);
     }
 }