+++ src/Resource/MyEnvelopeInfo.php
@@ -1,0 +1,17 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCz\DigiSign\Resource;
+
+class MyEnvelopeInfo extends BaseResource
+{
+    /** @var string */
+    public $id;
+
+    /** @var string */
+    public $accountId;
+
+    /** @var string */
+    public $accountName;
+}